remove decorator and fix bug with permissions for databases, extract status registry into its own class
