+++ src/main/resources/application.yaml
@@ -38,7 +38,6 @@ spring:
 
 mokey:
     cache:
-        enabled: true
         affiliations: paas, aurora
 
 logging:
